1_Array_Operations
An Array is a collection of similar / same elements. In this experiment the array can be 
represented as one / single dimensional elements.
Menu driven program in c - language to perform various array operations are implemented with 
the help of user defined functions as followings;
a. create()
b. display()
c. insert()
d. del ()
e. exit().